# perf(gossip): speed up handleBatchPullMessages