## Problem

The processMessages loop typically runs pretty fast, many times per second, until a certain point when it gets bottlenecked by an overwhelming number of push messages and is unable to catch up. The iteration duration grows exponentially until it takes minutes to deal with tens of thousands of messages. This continues for several minutes, until sig reaches a degenerate state where it is no longer receiving messages. The pull request thread is unable to identify any peers. This continues indefinitely. I have observed it continue for hours.

@0xNineteen confirmed that push messages are taking significantly longer to process than other messages:

![image](https://github.com/Syndica/sig/assets/15116250/b8d60b8a-ee1b-45ee-8d44-b82b393e7b3d)

## Cause

I timed individual pieces of handleBatchPushMessages and found that just the log messages alone accounted for around 70% of the execution time. Next in line was GossipTable.insertValues, which uses about 80-90% of the remaining time.

I measured the time of individual insertions, and it would rapidly increase to around 200-300 μs within the first minute of starting sig. At the time of a bottleneck, I'm seeing insertions take upwards of 500 μs.

In GossipTable.insertValues, there are ArrayList allocations in that function to return the result of the insertion. Completely eliminating that return and all of its allocations reduced the execution time to 70 μs. This is not an acceptable solution but it identifies the main contributing factor to its duration.

## Solution

First, I removed the log message. This eliminated the fatal bottlenecks. Instead it would reach recoverable bottlenecks. Each batch's processing time would usually be on the order of several seconds instead of exponentially growing numbers of minutes. After a couple minutes it would recover and go back to processing multiple batches per second.

To optimize GossipTable.insertValues, I pre-allocate an ArrayList to re-use for the failures every time it is called, and pass it in as a pointer, rather than allocating a new list for each insertion. This cut down the insertion time down to around over 120 μs.

After both changes there are no longer significant bottlenecks. processMessages always completes its processing in well under 1s.

I also moved the prune messages from info to debug so there are no per-message info-level logging. I replaced them with a summary info message. It still clutters the logs a lot, but it's an improvement.
```
time=2024-03-24T19:08:21Z level=info  gossip: recv 63 messages: 0 ping, 0 pong, 62 push, 1 pull request, 0 pull response, 0 prune
```


## Caveats

- Most of my initial benchmarking was done in Debug mode, but I realized my mistake and re-ran some of it in ReleaseFast. I confirmed there is still a problem there, and my change still resolves the issue.
- After running for tens of minutes, I sometimes see a long-lasting slowdown in the processMessages loop. Instead of looping hundreds of times per second, it may get to a state where it's only running a few times per second. However it does not appear to ever reach a runaway bottleneck where it simply can't keep up. That issue is resolved. It always works its way through its messages at least as fast as they are coming in.
- This does not fully resolve the NoPeers error when generating pull requests. It makes an improvement, but that error still comes up occasionally. Somehow we're not getting enough ping responses from other nodes. Maybe we need to send more pings.